re-build architecture for compose hybrid cloud paradigm (Athernex) (k8s, compose, Docker, and Gradle)
