--- .dockerignore
@@ -1,7 +1,0 @@
-.idea
-*.iml
-.git/
-.gradle
-gradle
-out/
-Dockerfile
