--- Dockerfile.only-package
@@ -1,9 +1,0 @@
-FROM openjdk:8-jre-slim
-
-EXPOSE 8080
-
-RUN mkdir /app
-
-COPY build/libs/*.jar /app/spring-boot-application.jar
-
-ENTRYPOINT ["java", "-XX:+UnlockExperimentalVMOptions", "-XX:+UseCGroupMemoryLimitForHeap", "-Djava.security.egd=file:/dev/./urandom","-jar","/app/spring-boot-application.jar"]
