--- mrgi_portal.yml
@@ -1,20 +1,0 @@
-version: '1.0'
-stages:
-  - prepare
-  - build
-steps:
-  main_clone:
-    title: Cloning main repository...
-    stage: prepare
-    type: git-clone
-    repo: 'charlesderek/mrgi-gradle-scaffold'
-    revision: master
-    git: github
-  BuildingDockerImage:
-    title: Building Docker Image
-    stage: build
-    type: build
-    image_name: mrgi-gradle-scaffold
-    working_directory: ./
-    tag: 'multi-stage-gradle'
-    dockerfile: Dockerfile
