+++ springbootapp/.dockerignore
@@ -1,0 +1,7 @@
+.idea
+*.iml
+.git/
+.gradle
+gradle
+out/
+Dockerfile
