+++ springbootapp/mrgi_portal-package-only.yml
@@ -1,0 +1,34 @@
+version: '1.0'
+stages:
+  - prepare
+  - test
+  - package
+  - build
+steps:
+  main_clone:
+    title: Cloning main repository...
+    stage: prepare
+    type: git-clone
+    repo: 'charlesderek/mrgi-gradle-scaffold'
+    revision: master
+    git: github
+  MyUnitTests:
+    title: Compile/Unit test
+    stage: test
+    image: gradle:4.7.0-jdk8-alpine
+    commands:
+      - gradle test --no-daemon --build-cache --gradle-user-home=/mrgi_portal/volume/.gradle -Dmaven.repo.local=/mrgi_portal/volume/m2
+  BuildMyJar:
+    title: Packaging Jar file
+    stage: package
+    image: gradle:4.7.0-jdk8-alpine
+    commands:
+     - gradle build --no-daemon --build-cache --gradle-user-home=/mrgi_portal/volume/.gradle -Dmaven.repo.local=/mrgi_portal/volume/m2
+  MyAppDockerImage:
+    title: Building Docker Image
+    type: build
+    stage: build
+    image_name: mrgi-gradle-scaffold
+    working_directory: ./
+    tag: 'non-multi-stage-gradle'
+    dockerfile: Dockerfile.only-package
