+++ springbootapp/src/main/java/com/mrgi_portal/gradle_scaffold/GradleExampleApplication.java
@@ -1,0 +1,13 @@
+package com.mrgi_portal.gradle_scaffold;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class GradleExampleApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(GradleExampleApplication.class, args);
+	}
+
+}
