--- src/main/resources/static/index.html
@@ -1,12 +1,0 @@
-<!DOCTYPE HTML>
-<html>
-<head>
-    <title>Gradle Docker example</title>
-    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
-</head>
-<body>
-    <p>Hello. I am a Java project, compiled by Gradle running inside Docker</a></p>
-
-    <p>See a quick <a href="actuator/health">status report</a></p>
-</body>
-</html>
